# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Freelancers",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://api-freelancer.otterscompany.com"
        }
    ],
    "paths": {
...
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile",
                "description": "Returns profile data",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Get profile.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user",
                "description": "Update a profile data",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lastname",
                        "in": "query",
                        "description": "Lastname of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "photo",
                        "in": "query",
                        "description": "Photo of user",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update profile user.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/profile/change/password": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update password user",
                "description": "Update password of a user",
                "operationId": "updateProfilePassword",
                "parameters": [
                    {
                        "name": "current_password",
                        "in": "query",
                        "description": "Current password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "description": "New password of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update password.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/profile/change/email": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update email user",
                "description": "Update email of a user",
                "operationId": "updateProfileEmail",
                "parameters": [
                    {
                        "name": "current_email",
                        "in": "query",
                        "description": "Current email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "new_email",
                        "in": "query",
                        "description": "New email of user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update email.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Not authenticated."
                    },
                    "422": {
                        "description": "Data not valid."
                    },
                    "500": {
                        "description": "An error occurred during the process."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}